app/requirements.txt
pymongo==3.10.1
celery==5.0.2
mongoengine==0.20.0
psycopg2-binary==2.8.5
django-cors-headers==3.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
scrapy-splash==0.7.2
attrs==23.1.0
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
cfscrape==2.1.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
distlib==0.3.7
Django==3.1.3
dnspython==2.4.2
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
packaging==23.1
parsel==1.7.0
platformdirs==3.10.0
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3.post1
queuelib==1.6.2
requests==2.31.0
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
scrapy-splash==0.7.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.2.4
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==1.25.10
virtualenv==20.24.2
w3lib==2.1.2
zope.interface==6.0
PyYAML==5.4.1
channels==3.0.3
asgiref==3.3.4
djangorestframework-xml==2.0.0
django-apscheduler==0.5.1